bocchi menu & some ui improvements

+++ src/minecraft/cn/stars/reversal/ui/atmoic/mainmenu/impl/MainGui.java
@@ -31,7 +31,7 @@ public class MainGui extends AtomicGui {
     }
 
     @Override
-    public void drawIcon(int posX, int posY, int color) {
+    public void drawIcon(float posX, float posY, int color) {
         atomic32.drawString(icon, posX - 0.2, posY - 0.5, color);
     }
