+++ src/minecraft/cn/stars/reversal/ui/atmoic/mainmenu/impl/MultiPlayerGui.java
@@ -63,7 +63,7 @@ public class MultiPlayerGui extends AtomicGui {
     }
 
     @Override
-    public void drawIcon(int posX, int posY, int color) {
+    public void drawIcon(float posX, float posY, int color) {
         atomic24.drawString(icon, posX, posY + 0.5, color);
     }
