+++ src/minecraft/cn/stars/reversal/ui/atmoic/mainmenu/impl/ReversalSettingsGui.java
@@ -27,7 +27,7 @@ public class ReversalSettingsGui extends AtomicGui {
     }
 
     @Override
-    public void drawIcon(int posX, int posY, int color) {
+    public void drawIcon(float posX, float posY, int color) {
         atomic24.drawString(icon, posX + 2, posY + 0.5, color);
     }
